Add files via upload
Recently, we had our garage door controller fail. Unfortunately new controller was unobtanium and buying a whole new motor would be quite costly. I decided to try and build a controller based on UNO and a few lines of code.
The build is based on Arduino UNO with low side PWM speed control and two relays for switching. The motor has high stall current (aprox. 10A) so H bridge was not an option or would be difficult to implement.
Since the remote receiver was a separate module, I was able to reverse engineer the signal output and all existing remotes work nicely.
All in all it took a week to do. It is installed now and works fine but app needs some fine tuning. The following functionalities are implemented:
1. Door opening/ closing
2. Door initialisation (detects and stores open and closed positions in eprom)
3. Obstacle detection (automatically stops if blocked for 1/10th sec)
to do:
1. Travel sensors (min & max)
2. Optical sensor (disable movement if door is blocked)
3. Automatic closing after car passed.